requirements.txt
aiofiles==23.2.1
aiogram==2.25
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.2.0
Babel==2.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
magic-filter==1.0.12
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
yarl==1.9.4
